requirements.txt
Flask
Flask-Cors
requests
python-dotenv
Gunicorn